Move tshirt_params to dedicated header file.
Creating tshirt_params.h and moving tshirt::tshirt_params struct from
Tshirt.h to this new file.
